sidecar container is correctly running oras
need to figure out the logic for it to run before the
main container, and then right after completion next,
and then customize with more annotation params. Then
after likely can make the docs and possibly automated
builds, but need to wait until nov 1st as out of actions
minutes!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>