WIP to add entrypoint logic
The oras cache sidecar needs to know how to pull an artifact,
and then create an indicator to the application that it is ready.
The application needs to wait for the indicator, and then
run an entrypoint that wraps the command. Since I want to keep
things simple (and not navigate adding a config map for a webhook
!) I am starting with a wget of the scripts instead of that. We
will see if this is a good idea or not!

Signed-off-by: vsoch <vsoch@users.noreply.github.com>