start of support for job (#10)
* start of support for job

The webhook is simplest if we allow it to watch both Job and Pod, and then simply act on whatever annotations we find under each. For Job, the operator looks for annotations on the Job.Spec.Template.Spec (a PodTemplate -> PodSpec) and for the Pod it looks directly at the Pod.Spec (also of type PofSpec) and this allows us to share functionality between the two (could be different) flows. Using "containers" as an abstraction also does the same.

This is a WIP and likely to change as we work further and learn from it. 

Signed-off-by: vsoch <vsoch@users.noreply.github.com>